pyproject.toml
[project]
name = "nuggets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "fastapi[standard]>=0.115.11",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "psycopg>=3.2.5",
    "pydantic-settings>=2.8.1",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "starlette>=0.46.1",
]
